WIP: Break out imports, buffer decode error
